Added custom node matching implementation
This is a complete custom implemenation of the node matching algorithm. The
primary improvement however is in the approximate node matching. All the nodes
that are not matching with the exact node match are grouped up based on their
node type. Within these groups all possible combinations are evaluated based on
how much the nodes differ. And the combinations that differ the least are
accepted as our aproximate matches.