Update GraphMergeHelper to use FDiffHelper
Important: This change disables applying and reverting of any of the diff types.
This functionality will be added back in a later update.

This ensures that everything now uses the FDiffHelper instead of
FGraphDiffControl. Allowing us to remove all temporary UI elements. And make
use of the improved diffs when refining the UI