Fix multiple links with same type causes the wrong link to be removed
In the cases of ApplyDiff_LINK_REMOVED and RevertDiff_LINK_ADDED we only
compared the link target by name and type, and did not check the node they
belonged to.

This caused an issue that if a pin was linked to two of the same pins, we would
always remove the link with the first of them.